refactor: reorg playback context typing
