--- .idea/csv-plugin.xml
@@ -1,51 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project version="4">
-  <component name="CsvFileAttributes">
-    <option name="attributeMap">
-      <map>
-        <entry key="/src/components/MusicPlayback/components/PlaybackControl/TimeBar/__tests__/index.test.tsx">
-          <value>
-            <Attribute>
-              <option name="separator" value="," />
-            </Attribute>
-          </value>
-        </entry>
-        <entry key="/src/components/MusicPlayback/components/PlaybackControl/TimeBar/index.tsx">
-          <value>
-            <Attribute>
-              <option name="separator" value=":" />
-            </Attribute>
-          </value>
-        </entry>
-        <entry key="/src/contexts/SpotifyWebPlayback/__tests__/index.test.tsx">
-          <value>
-            <Attribute>
-              <option name="separator" value="," />
-            </Attribute>
-          </value>
-        </entry>
-        <entry key="/src/contexts/SpotifyWebPlayback/hooks/usePlaybackStateMachine.ts">
-          <value>
-            <Attribute>
-              <option name="separator" value=";" />
-            </Attribute>
-          </value>
-        </entry>
-        <entry key="/src/hooks/spotify/query/useAvailableDevices.ts">
-          <value>
-            <Attribute>
-              <option name="separator" value="," />
-            </Attribute>
-          </value>
-        </entry>
-        <entry key="/src/pages/index/components/Present/PresentSuggestAlbum.tsx">
-          <value>
-            <Attribute>
-              <option name="separator" value=";" />
-            </Attribute>
-          </value>
-        </entry>
-      </map>
-    </option>
-  </component>
-</project>
