+++ src/components/MusicPlayback/components/ExtraControl/PickConnectedDevice/index.tsx
@@ -3,8 +3,10 @@ import { useErrorHandler } from 'react-error-boundary';
 import { ArrowContainer, Popover } from 'react-tiny-popover';
 import styled, { useTheme } from 'styled-components';
 
-import { useSpotifyWebPlayback } from '../../../../../contexts/SpotifyWebPlayback';
-import { PlaybackType } from '../../../../../contexts/SpotifyWebPlayback/states/PlaybackState';
+import {
+  PlaybackType,
+  useSpotifyWebPlayback,
+} from '../../../../../contexts/SpotifyWebPlayback';
 import type { theme } from '../../../../../contexts/Theme';
 import { useAvailableDevices } from '../../../../../hooks/spotify/query/useAvailableDevices';
 import { useToggle } from '../../../../../hooks/utils/useToggle';
