+++ src/components/MusicPlayback/components/ExtraControl/VolumeBar/__tests__/index.test.tsx
@@ -1,7 +1,7 @@
 import { render, screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 
-import { PlaybackType } from '../../../../../../contexts/SpotifyWebPlayback/states/PlaybackState';
+import { PlaybackType } from '../../../../../../contexts/SpotifyWebPlayback';
 import { AppThemeProvider } from '../../../../../../contexts/Theme';
 import { VolumeBar } from '../';
