+++ src/components/RemotePlaybackBanner/index.tsx
@@ -1,9 +1,11 @@
 import type { PropsWithChildren } from 'react';
 import styled from 'styled-components';
 
-import { useSpotifyWebPlayback } from '../../contexts/SpotifyWebPlayback';
-import type { PlaybackDevice } from '../../contexts/SpotifyWebPlayback/states/PlaybackState';
-import { PlaybackType } from '../../contexts/SpotifyWebPlayback/states/PlaybackState';
+import {
+  PlaybackType,
+  useSpotifyWebPlayback,
+} from '../../contexts/SpotifyWebPlayback';
+import type { PlaybackDevice } from '../../contexts/SpotifyWebPlayback/typings/Playback';
 import { ReactComponent as ConnectIcon } from './spotify-connect-icon.svg';
 
 const RemotePlaybackBannerContainer = styled.aside`
