+++ src/contexts/SpotifyWebPlayback/states/IdlePlaybackState.ts
@@ -4,8 +4,8 @@ import {
   ActivePlaybackState,
   PlaybackState,
   PlaybackType,
-  StateMachine,
-} from './PlaybackState';
+} from '../typings/Playback';
+import type { StateMachine } from '../typings/State';
 import { RepeatMode } from './RepeatMode';
 
 export class IdlePlaybackState implements ActivePlaybackState {
