+++ src/contexts/SpotifyWebPlayback/typings/Playback.ts
@@ -1,0 +1,43 @@
+import type { RepeatMode } from '../states/RepeatMode';
+import type { StateMachine } from './State';
+
+export enum PlaybackState {
+  IDLE = 'idle', // No playback taking control
+  INIT = 'init', // default state
+  PLAY_ON_LOCAL_PLAYBACK = 'playOnLocalPlayback', // track playing on local playback
+  PLAY_ON_REMOTE_PLAYBACK = 'playOnRemotePlayback', // track playing on remote playback
+}
+
+export enum PlaybackType {
+  Local = 'local',
+  Remote = 'remote',
+}
+
+export interface PlaybackDevice {
+  id: string;
+  is_active: boolean;
+  is_private_session: boolean;
+  is_restricted: boolean;
+  name: string;
+  type: string;
+  volume_percent: number;
+}
+
+export interface ActivePlaybackState {
+  readonly playbackType: PlaybackType;
+  readonly refreshInterval: number;
+  readonly stateMachine: StateMachine;
+  // eslint-disable-next-line typescript-sort-keys/interface
+  getPlaybackState(): Promise<null | {
+    actions: {
+      disallows: Spotify.PlaybackDisallows;
+    };
+    device: PlaybackDevice;
+    is_active: boolean;
+    is_paused: boolean;
+    progress_ms: number;
+    repeat_state: RepeatMode;
+    shuffle_state: boolean;
+    track: Spotify.Track;
+  }>;
+}
